Split a job on travis-ci
We often reach the travis-ci's limit which is 50 mins.
On the 'release' job we execute the headless.py script,
which creates a release build with features then run the reftest.
After that creates a release build without any feature set.

Split it into two jobs, one will execute the headless.py
and one will create a release build without features.